requirements.txt
Flask==1.1.1
requests==2.22.0
lxml==4.5.0
Flask-Caching==1.8.0
flask-swagger-ui==3.25.0
pyyaml==5.3.1
gunicorn==20.0.4
beautifulsoup4==4.9.0
html5lib==1.0.1
